Added a new folder "NutrReq4HouseholdEx" for use as RHS constraints
In this folder, there are five rds files and two csv files. The file names should be clear in terms of whether each dataset contains a lower bound or upper bound or acceptable range. I only included rows of demographic groups that are applicable to our household examples. The "Group" and "Age" combinations are consistent with the set up in the household examples. I added four count variables "house_1", "house_2" , "house_3", and "house_4" to indicate the number of members falling into a Group + Age row in each of the four household examples.